# docs/swagger.yaml
definitions:
  controllers.HttpError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  models.Student:
    properties:
      cpf:
        example: "01234567890"
        type: string
      name:
        example: Aluno Teste
        type: string
      rg:
        example: "123456789"
        type: string
    required:
    - cpf
    - name
    - rg
    type: object
  models.StudentView:
    properties:
      cpf:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      rg:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/students:
    get:
      consumes:
      - application/json
      description: Rota para listar todos os alunos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.StudentView'
            type: array
      summary: Listar alunos
      tags:
      - students
    post:
      consumes:
      - application/json
      description: Rota para cadastrar aluno
      parameters:
      - description: estrutura de aluno
        in: body
        name: aluno
        required: true
        schema:
          $ref: '#/definitions/models.Student'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.StudentView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.HttpError'
      summary: Cadastrar aluno
      tags:
      - students
  /api/v1/students/{id}:
    delete:
      consumes:
      - application/json
      description: Rota para deletar aluno
      parameters:
      - description: ID do aluno
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Deletar aluno
      tags:
      - students
    get:
      consumes:
      - application/json
      description: Rota para buscar aluno por ID
      parameters:
      - description: ID para buscar aluno
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StudentView'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.HttpError'
      summary: Buscar aluno por ID
      tags:
      - students
    patch:
      consumes:
      - application/json
      description: Rota para atualizar aluno
      parameters:
      - description: estrutura de aluno
        in: body
        name: aluno
        required: true
        schema:
          $ref: '#/definitions/models.Student'
      - description: ID do aluno
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StudentView'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.HttpError'
      summary: Atualizar aluno
      tags:
      - students
  /api/v1/students/cpf/{cpf}:
    get:
      consumes:
      - application/json
      description: Rota para buscar aluno por CPF
      parameters:
      - description: CPF para buscar aluno
        in: path
        name: cpf
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StudentView'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.HttpError'
      summary: Buscar aluno por cpf
      tags:
      - students
swagger: "2.0"